+++ uefi/src/fs/path/mod.rs
@@ -24,13 +24,12 @@ pub use path::{Components, Path};
 pub use pathbuf::PathBuf;
 
 use crate::data_types::chars::NUL_16;
-use crate::{CStr16, Char16, cstr16};
+use crate::{CStr16, Char16, char16, cstr16};
 pub use validation::PathError;
 pub(super) use validation::validate_path;
 
 /// The default separator for paths.
-// SAFETY: The memory is valid.
-pub const SEPARATOR: Char16 = unsafe { Char16::from_u16_unchecked('\\' as u16) };
+pub const SEPARATOR: Char16 = char16!('\\');
 
 /// Stringified version of [`SEPARATOR`].
 pub const SEPARATOR_STR: &CStr16 = cstr16!("\\");
@@ -38,18 +37,17 @@ pub const SEPARATOR_STR: &CStr16 = cstr16!("\\");
 /// Deny list of characters for path components. UEFI supports FAT-like file
 /// systems. According to <https://en.wikipedia.org/wiki/Comparison_of_file_systems>,
 /// paths should not contain these symbols.
-// SAFETY: The memory is valid.
-pub const CHARACTER_DENY_LIST: [Char16; 10] = unsafe {
+pub const CHARACTER_DENY_LIST: [Char16; 10] = {
     [
         NUL_16,
-        Char16::from_u16_unchecked('"' as u16),
-        Char16::from_u16_unchecked('*' as u16),
-        Char16::from_u16_unchecked('/' as u16),
-        Char16::from_u16_unchecked(':' as u16),
-        Char16::from_u16_unchecked('<' as u16),
-        Char16::from_u16_unchecked('>' as u16),
-        Char16::from_u16_unchecked('?' as u16),
+        char16!('"'),
+        char16!('*'),
+        char16!('/'),
+        char16!(':'),
+        char16!('<'),
+        char16!('>'),
+        char16!('?'),
         SEPARATOR,
-        Char16::from_u16_unchecked('|' as u16),
+        char16!('|'),
     ]
 };